+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 * Updated dependencies
 * Upgraded gradle to 8.13.2
 * Updated documentation thanks [@makoConstruct](https://github.com/makoConstruct) ([Issue #1](https://github.com/NexusDynamic/flutter_refresh_rate_control/issues/1))
-
+* Added documentation for the `getRefreshRateInfo()` method
+* Added default keys for the `getRefreshRateInfo()` that are always returned.
+  - `platform`: (string) The platform the app is running on, e.g. "android", "ios", "windows", "linux", "macos"
+  - `supported`: (bool) Whether refresh rate control is supported on the current platform
 # 0.0.4+1
 
 * Lower Dart SDK minimum version to `3.7.0`